pyproject.toml
[project]
name = "keygene"
version = "0.2.0"
description = "Code for the KeyGene project at YES!Delft"
authors = [
    { name = "Romir Kulshrestha", email = "r.kulshrestha@student.tudelft.nl" },
]
license = { text = "MIT License" }
readme = "README.md"
requires-python = ">=3.7, <3.8"
classifiers = [
    "Programming Language :: Python :: 3.7",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Development Status :: 3 - Alpha",
]
dependencies = [
    "Pillow",
    "bosdyn-client >= 3.1",
    "bosdyn-mission >= 3.1",
    "grpcio == 1.43.0",
    "opencv-python",
    "protobuf ~= 3.19.4",
    "qrcode",
    "tqdm",
    "windows-curses ; sys_platform == 'win32'",
    "nicegui",
    "python-socketio<5.10.0"
]

[project.urls]
Repository = "https://github.com/romirk/spot-keygene"
Home-Page = "https://romir.eu/"

[project.scripts]
kg = "keygene.__main__:main"
